Crate pallet_loans

Loans pallet

Overview

Loans pallet implement the lending protocol by using a pool-based strategy that aggregates each user’s supplied assets. The interest rate is dynamically determined by the supply and demand.

Re-exports

pub use pallet::*;

pub use weights::WeightInfo;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights
Autogenerated weights for pallet_loans

Structs

BorrowSnapshot
Container for borrow balance information

CurveModel
The curve interest rate model

Deposits
Deposit information

EarnedSnapshot
Container for earned amount information

JumpModel
The jump interest rate model

Market
Market.

Enums

InterestRateModel
Parallel interest rate model

MarketState
The current state of a market. For more information, see Market.

Functions

accrued_interest

increment_index