Module pallet_loans::math

Module provides all the interest and rate related calculations

Enums

Adjustment

represents how much some other value should be incremented or decremented by

Functions

bullet_loan_present_value

calculates present value for bullet loan assumes maturity date has not passed https://docs.centrifuge.io/learn/pool-valuation/#simple-example-for-one-financing

bullet_loan_risk_adjusted_expected_cash_flow

calculates the risk adjusted expected cash flow for bullet loan type assumes maturity date has not passed https://docs.centrifuge.io/learn/pool-valuation/#simple-example-for-one-financing

calculate_accumulated_rate

calculates the latest accumulated rate since the last

calculate_normalized_debt

calculates the principal debt after the adjustment current_debt and cumulative_rate must be latest

debt

calculates the debt using debt=principal_debt * cumulative_rate

interest_rate_per_sec

calculates rate per second from the given nominal interest rate https://docs.centrifuge.io/learn/interest-rate-methodology/

penalty_interest_rate_per_sec

calculates fractional part of the rate per second from the given nominal interest rate https://docs.centrifuge.io/learn/interest-rate-methodology/