Module pallet_loans::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

GenesisConfig
Can be used to configure the genesis state of this pallet.

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageAccountBorrows

_GeneratedPrefixForStorageAccountDeposits

_GeneratedPrefixForStorageAccountEarned

_GeneratedPrefixForStorageBorrowIndex

_GeneratedPrefixForStorageBorrowRate

_GeneratedPrefixForStorageExchangeRate

_GeneratedPrefixForStorageLastBlockTimestamp

_GeneratedPrefixForStorageMarkets

_GeneratedPrefixForStorageSupplyRate

_GeneratedPrefixForStorageTotalBorrows

_GeneratedPrefixForStorageTotalReserves

_GeneratedPrefixForStorageTotalSupply

_GeneratedPrefixForStorageUtilizationRatio

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

Traits

Config
Configuration trait of this pallet.

Type Definitions

AccountBorrows
Mapping of account addresses to outstanding borrow balances CurrencyType -> Owner -> BorrowSnapshot

AccountDeposits
Mapping of account addresses to deposit details CollateralType -> Owner -> Deposits

AccountEarned
Mapping of account addresses to total deposit interest accrual CurrencyType -> Owner -> EarnedSnapshot

BorrowIndex
Accumulator of the total earned interest rate since the opening of the market CurrencyType -> u128

BorrowRate
Mapping of borrow rate to currency type

ExchangeRate
The exchange rate from the underlying to the internal collateral

LastBlockTimestamp
The timestamp of the previous block or defaults to timestamp at genesis.

Markets
Mapping of currency id to its market

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

SupplyRate
Mapping of supply rate to currency type

TotalBorrows
Total amount of outstanding borrows of the underlying in this market CurrencyType -> Balance

TotalReserves
Total amount of reserves of the underlying held in this market CurrencyType -> Balance

TotalSupply
Total number of collateral tokens in circulation CollateralType -> Balance

UtilizationRatio
Borrow utilization ratio