Struct pallet_loans::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn active_market(
origin: OriginFor<T>,
currency_id: CurrencyId
) -> DispatchResultWithPostInfo

Activates a market. Returns Err if the market currency does not exist.

If the market is already activated, does nothing.

currency_id: Market related currency

pub fn add_market(
origin: OriginFor<T>,
currency_id: CurrencyId,
market: Market
) -> DispatchResultWithPostInfo

Stores a new market and its related currency. Returns Err if a currency is not attached to an existent market.

All provided market states must be Pending, otherwise an error will be returned.

If a currency is already attached to a market, then the market will be replaced by the new provided value.

currency_id: Market related currency

market: The market that is going to be stored

pub fn update_market(
origin: OriginFor<T>,
currency_id: CurrencyId,
market: Market
) -> DispatchResultWithPostInfo

Updates a stored market. Returns Err if the market currency does not exist.

Market state won’t be modified, regardless of the provided value.

currency_id: Market related currency

market: The new market parameters

pub fn mint(
origin: OriginFor<T>,
currency_id: CurrencyId,
mint_amount: Balance
) -> DispatchResultWithPostInfo

Sender supplies assets into the market and receives internal supplies in exchange.

currency_id: the asset to be deposited.

mint_amount: the amount to be deposited.

pub fn redeem(
origin: OriginFor<T>,
currency_id: CurrencyId,
redeem_amount: Balance
) -> DispatchResultWithPostInfo

Sender redeems some of internal supplies in exchange for the underlying asset.

currency_id: the asset to be redeemed.

redeem_amount: the amount to be redeemed.

pub fn redeem_all(
origin: OriginFor<T>,
currency_id: CurrencyId
) -> DispatchResultWithPostInfo

Sender redeems all of internal supplies in exchange for the underlying asset.

currency_id: the asset to be redeemed.

pub fn borrow(
origin: OriginFor<T>,
currency_id: CurrencyId,
borrow_amount: Balance
) -> DispatchResultWithPostInfo

Sender borrows assets from the protocol to their own address.

currency_id: the asset to be borrowed.

borrow_amount: the amount to be borrowed.

pub fn repay_borrow(
origin: OriginFor<T>,
currency_id: CurrencyId,
repay_amount: Balance
) -> DispatchResultWithPostInfo

Sender repays some of their debts.

currency_id: the asset to be repaid.

repay_amount: the amount to be repaid.

pub fn repay_borrow_all(
origin: OriginFor<T>,
currency_id: CurrencyId
) -> DispatchResultWithPostInfo

Sender repays all of their debts.

currency_id: the asset to be repaid.

pub fn transfer_token(
origin: OriginFor<T>,
to: T::AccountId,
currency_id: CurrencyId,
amount: Balance
) -> DispatchResultWithPostInfo

Using for development

pub fn collateral_asset(
origin: OriginFor<T>,
currency_id: CurrencyId,
enable: bool
) -> DispatchResultWithPostInfo

Set the collateral asset.

currency_id: the asset to be set.

enable: turn on/off the collateral option.

pub fn liquidate_borrow(
origin: OriginFor<T>,
borrower: T::AccountId,
liquidate_token: CurrencyId,
repay_amount: Balance,
collateral_token: CurrencyId
) -> DispatchResultWithPostInfo

The sender liquidates the borrower’s collateral.

borrower: the borrower to be liquidated.

liquidate_token: the assert to be liquidated.

repay_amount: the amount to be repaid borrow.

collateral_token: The collateral to seize from the borrower.

pub fn add_reserves(
origin: OriginFor<T>,
payer: <T::Lookup as StaticLookup>::Source,
currency_id: CurrencyId,
add_amount: Balance
) -> DispatchResultWithPostInfo

Add reserves by transferring from payer.

May only be called from T::ReserveOrigin.

payer: the payer account.

currency_id: the assets to be added.

add_amount: the amount to be added.

pub fn reduce_reserves(
origin: OriginFor<T>,
receiver: <T::Lookup as StaticLookup>::Source,
currency_id: CurrencyId,
reduce_amount: Balance
) -> DispatchResultWithPostInfo

Reduces reserves by transferring to receiver.

May only be called from T::ReserveOrigin.

receiver: the receiver account.

currency_id: the assets to be reduced.

reduce_amount: the amount to be reduced.

impl<T: Config> Pallet<T>

pub fn last_block_timestamp() -> Timestamp

The timestamp of the previous block or defaults to timestamp at genesis.

impl<T: Config> Pallet<T>

pub fn total_supply<KArg>(k: KArg) -> Balance where
KArg: EncodeLike<CurrencyId>,

Total number of collateral tokens in circulation CollateralType -> Balance

impl<T: Config> Pallet<T>

pub fn total_borrows<KArg>(k: KArg) -> Balance where
KArg: EncodeLike<CurrencyId>,

Total amount of outstanding borrows of the underlying in this market CurrencyType -> Balance

impl<T: Config> Pallet<T>

pub fn total_reserves<KArg>(k: KArg) -> Balance where
KArg: EncodeLike<CurrencyId>,

Total amount of reserves of the underlying held in this market CurrencyType -> Balance

impl<T: Config> Pallet<T>

pub fn account_borrows<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> BorrowSnapshot where
KArg1: EncodeLike<CurrencyId>,
KArg2: EncodeLike<T::AccountId>,

Mapping of account addresses to outstanding borrow balances CurrencyType -> Owner -> BorrowSnapshot

impl<T: Config> Pallet<T>

pub fn account_deposits<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Deposits where
KArg1: EncodeLike<CurrencyId>,
KArg2: EncodeLike<T::AccountId>,

Mapping of account addresses to deposit details CollateralType -> Owner -> Deposits

impl<T: Config> Pallet<T>

pub fn account_earned<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> EarnedSnapshot where
KArg1: EncodeLike<CurrencyId>,
KArg2: EncodeLike<T::AccountId>,

Mapping of account addresses to total deposit interest accrual CurrencyType -> Owner -> EarnedSnapshot

impl<T: Config> Pallet<T>

pub fn borrow_index<KArg>(k: KArg) -> Rate where
KArg: EncodeLike<CurrencyId>,

Accumulator of the total earned interest rate since the opening of the market CurrencyType -> u128

impl<T: Config> Pallet<T>

pub fn exchange_rate<KArg>(k: KArg) -> Rate where
KArg: EncodeLike<CurrencyId>,

The exchange rate from the underlying to the internal collateral

impl<T: Config> Pallet<T>

pub fn borrow_rate<KArg>(k: KArg) -> Rate where
KArg: EncodeLike<CurrencyId>,

Mapping of borrow rate to currency type

impl<T: Config> Pallet<T>

pub fn supply_rate<KArg>(k: KArg) -> Rate where
KArg: EncodeLike<CurrencyId>,

Mapping of supply rate to currency type

impl<T: Config> Pallet<T>

pub fn utilization_ratio<KArg>(k: KArg) -> Ratio where
KArg: EncodeLike<CurrencyId>,

Borrow utilization ratio

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

pub fn get_account_liquidity(
account: &T::AccountId
) -> Result<(Liquidity, Shortfall), DispatchError>

pub fn redeem_internal(
who: &T::AccountId,
currency_id: &CurrencyId,
voucher_amount: Balance
) -> Result<Balance, DispatchError>

pub fn current_borrow_balance(
who: &T::AccountId,
currency_id: &CurrencyId
) -> Result<Balance, DispatchError>

pub fn current_balance_from_snapshot(
currency_id: &CurrencyId,
snapshot: BorrowSnapshot
) -> Result<Balance, DispatchError>

Same as current_borrow_balance but takes a given snapshot instead of fetching the storage

pub fn liquidate_borrow_internal(
liquidator: T::AccountId,
borrower: T::AccountId,
liquidate_currency_id: CurrencyId,
repay_amount: Balance,
collateral_currency_id: CurrencyId
) -> DispatchResult

Note:

liquidate_token is borrower’s debt asset.

collateral_token is borrower’s collateral asset.

repay_amount is amount of liquidate_token

The liquidator will repay a certain amount of liquidate_token from own account for borrower. Then the protocol will reduce borrower’s debt and liquidator will receive collateral_token(as voucher amount) from borrower.

pub fn update_exchange_rate(currency_id: CurrencyId) -> DispatchResult

pub fn calc_utilization_ratio(
cash: Balance,
borrows: Balance,
reserves: Balance
) -> Result<Ratio, DispatchError>

pub fn update_borrow_index(
borrow_rate: Rate,
currency_id: CurrencyId,
market: &Market
) -> DispatchResult

pub fn get_total_cash(currency_id: CurrencyId) -> Balance

pub fn calc_underlying_amount(
voucher_amount: u128,
exchange_rate: Rate
) -> Result<Balance, DispatchError>

pub fn calc_collateral_amount(
underlying_amount: u128,
exchange_rate: Rate
) -> Result<Balance, DispatchError>

pub fn get_price(currency_id: &CurrencyId) -> Result<Price, Error<T>>

pub fn market(currency_id: &CurrencyId) -> Result<Market, DispatchError>