Crate pallet_loans

This pallet offers extrinsics to handle loans.

The following actions are performed over a loan:

Extrinsics	Role

Pallet::create()	Borrower

Pallet::borrow()	Borrower

Pallet::repay()	Borrower

Pallet::write_off()

Pallet::admin_write_off()	LoanAdmin

Pallet::propose_loan_mutation()	LoanAdmin

Pallet::apply_loan_mutation()

Pallet::propose_transfer_debt()	Borrower

Pallet::apply_transfer_debt()

Pallet::close()	Borrower

The following actions are performed over an entire pool of loans:

Extrinsics	Role

Pallet::propose_write_off_policy()	PoolAdmin

Pallet::apply_write_off_policy()

Pallet::update_portfolio_valuation()

The whole pallet is optimized for the more expensive extrinsic that is Pallet::update_portfolio_valuation() that should go through all active loans.

Re-exports

pub use pallet::*;

Modules

entities

High level types that uses pallet::Config

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

Low level types that doesn’t know about what a pallet is Contains base types without Config references

util

Utility types for configure the pallet from a runtime

Traits

WeightInfo