Crate pallet_loans

Loan pallet

This pallet provides functionality for managing loans on Tinlake

Re-exports

pub use pallet::*;

Modules

functions

Module provides loan related functions

loan_type

Module provides functionality for different loan types

math

Module provides all the interest and rate related calculations

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

Module provides base types and their functions

weights

Autogenerated weights for pallet_loans

Enums

Adjustment

PermissionScope

PoolRole

PoolRole can hold any type of role specific functions a user can do on a given pool.

Role

The Role enum is used by the permissions pallet, to specify which role an account has within a specific scope.