Trait pallet_loans::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeChange: From<Change<Self>> + TryInto<Change<Self>>;
    type CurrencyId: Parameter + Copy + MaxEncodedLen;
    type CollectionId: Parameter + Member + Default + TypeInfo + Copy + MaxEncodedLen;
    type ItemId: Parameter + Member + Default + TypeInfo + Copy + MaxEncodedLen;
    type LoanId: Parameter + Member + Default + TypeInfo + MaxEncodedLen + Copy + EnsureAdd + One;
    type PriceId: Parameter + Member + TypeInfo + Copy + MaxEncodedLen;
    type Rate: Parameter + Member + FixedPointNumber + TypeInfo + MaxEncodedLen;
    type Balance: Balance + FixedPointOperand;
    type Quantity: Parameter + Member + FixedPointNumber + TypeInfo + MaxEncodedLen;
    type PerThing: Parameter + Member + PerThing + TypeInfo + MaxEncodedLen;
    type Time: TimeAsSecs;
    type Moment: Parameter + Member + IntoSeconds;
    type NonFungible: Transfer<Self::AccountId> + Inspect<Self::AccountId, CollectionId = Self::CollectionId, ItemId = Self::ItemId>;
    type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen;
    type Pool: PoolReserve<Self::AccountId, Self::CurrencyId, Balance = Self::Balance, PoolId = Self::PoolId>;
    type Permissions: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role, Error = DispatchError>;
    type PriceRegistry: DataRegistry<Self::PriceId, Self::PoolId, Data = PriceOf<Self>>;
    type InterestAccrual: InterestAccrual<Self::Rate, Self::Balance, Adjustment<Self::Balance>, NormalizedDebt = Self::Balance>;
    type ChangeGuard: ChangeGuard<PoolId = Self::PoolId, ChangeId = Self::Hash, Change = Self::RuntimeChange>;
    type MaxActiveLoansPerPool: Get<u32>;
    type MaxWriteOffPolicySize: Get<u32> + Parameter;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type RuntimeChange: From<Change<Self>> + TryInto<Change<Self>>

Represent a runtime change

type CurrencyId: Parameter + Copy + MaxEncodedLen

Identify a curreny.

type CollectionId: Parameter + Member + Default + TypeInfo + Copy + MaxEncodedLen

Identify a non fungible collection

type ItemId: Parameter + Member + Default + TypeInfo + Copy + MaxEncodedLen

Identify a non fungible item

type LoanId: Parameter + Member + Default + TypeInfo + MaxEncodedLen + Copy + EnsureAdd + One

Identify a loan in the pallet

type PriceId: Parameter + Member + TypeInfo + Copy + MaxEncodedLen

Identify a loan in the pallet

type Rate: Parameter + Member + FixedPointNumber + TypeInfo + MaxEncodedLen

Defines the rate type used for math computations

type Balance: Balance + FixedPointOperand

Defines the balance type used for math computations

type Quantity: Parameter + Member + FixedPointNumber + TypeInfo + MaxEncodedLen

Type to represent different quantities

type PerThing: Parameter + Member + PerThing + TypeInfo + MaxEncodedLen

Defines the perthing type used where values can not overpass 100%

type Time: TimeAsSecs

Fetching method for the time of the current block

type Moment: Parameter + Member + IntoSeconds

Generic time type

type NonFungible: Transfer<Self::AccountId> + Inspect<Self::AccountId, CollectionId = Self::CollectionId, ItemId = Self::ItemId>

Used to mint, transfer, and inspect assets.

type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen

The PoolId type

type Pool: PoolReserve<Self::AccountId, Self::CurrencyId, Balance = Self::Balance, PoolId = Self::PoolId>

Access to the pool

type Permissions: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role, Error = DispatchError>

Used to verify permissions of users

type PriceRegistry: DataRegistry<Self::PriceId, Self::PoolId, Data = PriceOf<Self>>

Used to fetch and update Oracle prices

type InterestAccrual: InterestAccrual<Self::Rate, Self::Balance, Adjustment<Self::Balance>, NormalizedDebt = Self::Balance>

Used to calculate interest accrual for debt.

type ChangeGuard: ChangeGuard<PoolId = Self::PoolId, ChangeId = Self::Hash, Change = Self::RuntimeChange>

Used to notify the runtime about changes that require special treatment.

type MaxActiveLoansPerPool: Get<u32>

Max number of active loans per pool.

type MaxWriteOffPolicySize: Get<u32> + Parameter

Max number of write-off groups per pool.

type WeightInfo: WeightInfo

Information of runtime weights