Trait pallet_loans::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type ClassId: Parameter + Member + MaybeSerializeDeserialize + Copy + Default + TypeInfo + IsType<<<Self as Config>::NonFungible as Inspect<<Self as Config>::AccountId>>::CollectionId>;
    type LoanId: Parameter + Member + MaybeSerializeDeserialize + Copy + TypeInfo + From<u128> + IsType<<<Self as Config>::NonFungible as Inspect<<Self as Config>::AccountId>>::ItemId>;
    type Rate: Parameter + Member + MaybeSerializeDeserialize + FixedPointNumber + TypeInfo;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>;
    type NonFungible: Transfer<Self::AccountId> + Mutate<Self::AccountId>;
    type Time: UnixTime;
    type LoansPalletId: Get<PalletId>;
    type Pool: PoolReserve<Self::AccountId, Self::CurrencyId, Balance = Self::Balance>;
    type CurrencyId: Parameter + Copy;
    type Permission: PermissionsT<Self::AccountId, Scope = PermissionScope<<<Self as Config>::Pool as PoolInspect<<Self as Config>::AccountId, <Self as Config>::CurrencyId>>::PoolId, Self::CurrencyId>, Role = Role, Error = DispatchError>;
    type InterestAccrual: InterestAccrualT<Self::Rate, Self::Balance, Adjustment<Self::Balance>>;
    type WeightInfo: WeightInfo;
    type MaxActiveLoansPerPool: Get<u32>;
    type MaxWriteOffGroups: Get<u32>;
    type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type ClassId: Parameter + Member + MaybeSerializeDeserialize + Copy + Default + TypeInfo + IsType<<<Self as Config>::NonFungible as Inspect<<Self as Config>::AccountId>>::CollectionId>

The ClassId type

type LoanId: Parameter + Member + MaybeSerializeDeserialize + Copy + TypeInfo + From<u128> + IsType<<<Self as Config>::NonFungible as Inspect<<Self as Config>::AccountId>>::ItemId>

The LoanId/InstanceId type

type Rate: Parameter + Member + MaybeSerializeDeserialize + FixedPointNumber + TypeInfo

the rate type

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>

type NonFungible: Transfer<Self::AccountId> + Mutate<Self::AccountId>

The NonFungible trait that can mint, transfer, and inspect assets.

type Time: UnixTime

A way for use to fetch the time of the current block

type LoansPalletId: Get<PalletId>

PalletID of this loan module

type Pool: PoolReserve<Self::AccountId, Self::CurrencyId, Balance = Self::Balance>

Pool reserve type

type CurrencyId: Parameter + Copy

type Permission: PermissionsT<Self::AccountId, Scope = PermissionScope<<<Self as Config>::Pool as PoolInspect<<Self as Config>::AccountId, <Self as Config>::CurrencyId>>::PoolId, Self::CurrencyId>, Role = Role, Error = DispatchError>

Permission type that verifies permissions of users

type InterestAccrual: InterestAccrualT<Self::Rate, Self::Balance, Adjustment<Self::Balance>>

type WeightInfo: WeightInfo

Weight info trait for extrinsics

type MaxActiveLoansPerPool: Get<u32>

Max number of active loans per pool.

type MaxWriteOffGroups: Get<u32>

Max number of write-off groups per pool.

type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>

Source of the current block number