Trait pallet_loans::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type PriceFeeder: PriceFeeder;
    type PalletId: Get<PalletId>;
    type ReserveOrigin: EnsureOrigin<Self::Origin>;
    type UpdateOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;
    type UnixTime: UnixTime;
    type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
    type RewardAssetId: Get<<<Self as Config>::Assets as Inspect<<Self as Config>::AccountId>>::AssetId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type PriceFeeder: PriceFeeder

The oracle price feeder

type PalletId: Get<PalletId>

The loan’s module id, keep all collaterals of CDPs.

type ReserveOrigin: EnsureOrigin<Self::Origin>

The origin which can add/reduce reserves.

type UpdateOrigin: EnsureOrigin<Self::Origin>

The origin which can update rate model, liquidate incentive and add/reduce reserves. Root can always do this.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type UnixTime: UnixTime

Unix time

type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Assets for deposit/withdraw collateral assets to/from loans module

type RewardAssetId: Get<<<Self as Config>::Assets as Inspect<<Self as Config>::AccountId>>::AssetId>

Reward asset id.